Create Epipolar geometry
The code is not working although I have added the "#include <opencv2/opencv.hpp>", it gives some errors that are caused because it doesnt recognizes the functions that should be inside the opencv library. Im going away today for two weeks but when iIreturn I will try again to put the code running properly. If somebody could give a look at it and maybe solve the problem i would be very grateful. Any question: facebook or mail.
Catarina Lima